fix db error for simps
