+++ frameworks/PHP/simps/README.md
@@ -1,13 +1,13 @@
 <p align="center">
     <a href="https://simps.io" target="_blank">
-        <img src="https://cdn.jsdelivr.net/gh/sy-records/staticfile/images/simps.png" alt="Simps" />
+        <img src="https://cdn.jsdelivr.net/gh/sy-records/staticfile@master/images/simps.png" alt="Simps" />
     </a>
 </p>
 
-[![Simps License](https://img.shields.io/packagist/l/simple-swoole/simps?color=blue)](https://github.com/simple-swoole/simps/blob/master/LICENSE)
+[![Simps License](https://poser.pugx.org/simple-swoole/simps/license)](https://github.com/simple-swoole/simps/blob/master/LICENSE)
 [![Latest Version](https://img.shields.io/packagist/v/simple-swoole/simps.svg)](https://packagist.org/packages/simple-swoole/simps)
 [![Simps Doc](https://img.shields.io/badge/docs-passing-blue.svg)](https://doc.simps.io)
-[![Contact Simps Team](https://img.shields.io/badge/contact-@SimpsTeam-blue.svg?style=flat)](mailto:team@simps.io)
+[![Contact Simps Team](https://img.shields.io/badge/contact-@Simps%20Team-blue.svg)](mailto:team@simps.io)
 [![Php Version](https://img.shields.io/badge/php-%3E=7.1-brightgreen.svg)](https://www.php.net)
 [![Swoole Version](https://img.shields.io/badge/swoole-%3E=4.4.0-brightgreen.svg)](https://github.com/swoole/swoole-src)
 
@@ -18,7 +18,7 @@
 🚀 A simple, lightweight and high-performance PHP coroutine framework.
 
 [Home Page](https://simps.io)
-[Document](https://doc.simps.io)
+[Documentation](https://doc.simps.io)
 
 ## Important Libraries
